% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matcher.R
\name{blazingMatch}
\alias{blazingMatch}
\title{blazing fast propensity score matching}
\usage{
blazingMatch(
  data,
  treatment_var,
  confounding_vars,
  treatment_group,
  fast = FALSE
)
}
\arguments{
\item{data}{data frame containing data}

\item{treatment_var}{treatment variable name}

\item{confounding_vars}{confounding variable names}

\item{treatment_group}{value corresponding to the treatment group in treatment variable}
}
\value{
list containing two data frame `treatment` and `control`.
}
\description{
blazing fast propensity score matching
}
\examples{
data(lalonde)
confounding_vars <- c('age', 'educ', 'race', 'nodegree', 'married', 're74', 're75')
result <- blazingMatch(data=lalonde, treatment_var='treat', confounding_vars = confounding_vars, treatment_group = 1)
summary(result$treatment)
summary(result$control)
}
