% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matcher.R
\name{closest_index}
\alias{closest_index}
\title{quick nearest neighbor search for propensity scores}
\usage{
closest_index(ctr_score, trt_score)
}
\arguments{
\item{ctr_score}{data frame containing index and propensity scores for control group}

\item{trt_score}{data frame containing index and propensity scores for treatment group}
}
\description{
quick nearest neighbor search for propensity scores
}
