% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matcher.R
\name{closestIndex}
\alias{closestIndex}
\title{quick nearest neighbor search for propensity scores}
\usage{
closestIndex(ctr_score, trt_score)
}
\arguments{
\item{ctr_score}{data frame containing index and propensity scores for control group}

\item{trt_score}{data frame containing index and propensity scores for treatment group}
}
\value{
data frame
}
\description{
quick nearest neighbor search for propensity scores
}
