% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matcher.R
\name{closestIndexCpp}
\alias{closestIndexCpp}
\title{c++ wrapper function for quick nearest neighbor search for propensity scores}
\usage{
closestIndexCpp(ctr_score, trt_score)
}
\arguments{
\item{ctr_score}{data frame containing index and propensity scores for control group}

\item{trt_score}{data frame containing index and propensity scores for treatment group}
}
\value{
data frame
}
\description{
c++ wrapper function for quick nearest neighbor search for propensity scores
}
