% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nearestScore}
\alias{nearestScore}
\title{nearest score search c++ implementation.}
\usage{
nearestScore(ctr_score, trt_score)
}
\arguments{
\item{ctr_score}{matrix containing control group index and score}

\item{trt_score}{matrix containing treatment group index and score}
}
\value{
matched treatment and control indices
}
\description{
nearest score search c++ implementation.
}
\examples{

trt_score = data.frame(index=c(1:5), score=c(0.1, 0.03, 0.7, 0.2, 0.5))
ctr_score = data.frame(index=c(1:20), score=runif(20))
trt_score = trt_score[order(trt_score$score), ]
ctr_score = ctr_score[order(ctr_score$score), ]
nearestScore(as.matrix(ctr_score), as.matrix(trt_score))

}
